+++ examples/essa-engine/model-benchmark.cpp
@@ -17,24 +17,33 @@ constexpr size_t ModelCount = 100000;
 class WorldView : public GUI::Widget {
 public:
     WorldView()
-        : m_model(GUI::Application::the().resource_manager().require_external<Essa::Model>("../examples/essa-engine/tree.obj")) {
-    }
+        : m_model(GUI::Application::the()
+                      .resource_manager()
+                      .require_external<Essa::Model>(
+                          "../examples/essa-engine/tree.obj")) { }
 
 private:
     virtual void draw(Gfx::Painter& painter) const override {
         GUI::WorldDrawScope scope { painter };
 
-        llgl::Camera camera { llgl::Projection::perspective({ 1.4, raw_size().x() / raw_size().y(), 0.1, 30 }, {}) };
+        llgl::Camera camera { llgl::Projection::perspective(
+            { 1.4, static_cast<double>(raw_size().x() / raw_size().y()), 0.1,
+                30 },
+            {}) };
         camera = camera.translate({ 0, 2, 6 });
 
         Essa::Shaders::Lighting::Uniforms uniforms;
-        uniforms.set_transform({}, camera.view_matrix(), camera.projection().matrix());
+        uniforms.set_transform(
+            {}, camera.view_matrix(), camera.projection().matrix());
 
         for (size_t s = 0; s < ModelCount; s++) {
             auto uniforms_model = uniforms;
             auto rndx = Util::Random::floating<float>(-4, 4);
             auto rndy = Util::Random::floating<float>(-22, 2);
-            uniforms_model.set_model(llgl::Transform {}.translate({ rndx, 0, rndy }).scale(0.05).matrix());
+            uniforms_model.set_model(llgl::Transform {}
+                                         .translate({ rndx, 0, rndy })
+                                         .scale(0.05)
+                                         .matrix());
             m_model.render(painter.renderer(), m_shader, uniforms_model);
         }
     };
@@ -44,7 +53,9 @@ private:
 };
 
 int main() {
-    GUI::SimpleApplication<GUI::Container> app { "Essa::Model rendering benchmark" };
+    GUI::SimpleApplication<GUI::Container> app {
+        "Essa::Model rendering benchmark"
+    };
 
     auto& main_widget = app.main_widget();
     main_widget.set_layout<GUI::BasicLayout>();
@@ -55,9 +66,7 @@ int main() {
     auto label = main_widget.add_widget<GUI::Textfield>();
     label->set_size({ 400.0_px, 50.0_px });
 
-    app.on_tick = [&]() {
-        label->set_content(Util::to_ustring(app.tps()));
-    };
+    app.on_tick = [&]() { label->set_content(Util::to_ustring(app.tps())); };
 
     app.run();
     return 0;
